+++ lib/src/printer/pluggable.dart
@@ -3,7 +3,7 @@ import '../../printer.dart';
 /// Printer callback function.
 typedef PrintCallback = String Function(Object? value);
 
-/// Prints a string literal.
+/// Prints an object using a callback function.
 class PluggablePrinter extends Printer {
   final PrintCallback callback;